chore: rename fsUtils to fs-utils
