+++ scripts/fs-utils.js
